fix double

+++ train.py
@@ -80,7 +80,7 @@ def train_model(model, dataloaders, criterion, optimizer, num_epochs):
                     running_corrects += torch.sum(preds == labels.data)
 
             epoch_loss = running_loss / len(dataloaders[phase])
-            epoch_acc = running_corrects.double() / len(dataloaders[phase])
+            epoch_acc = running_corrects / len(dataloaders[phase])
 
             print("{} Loss: {:.4f} Acc: {:.4f}".format(phase, epoch_loss, epoch_acc))
